requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
flake8==4.0.1
greenlet==1.1.2
idna==3.3
IMDbPY==2021.4.18
isort==5.10.1
lxml==4.8.0
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.32
urllib3==1.26.9
